refactor(LottoController): 입출력 클래스 변경
- InputView, OutputView 클래스에서 object로 변경
