+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,33 +1,31 @@
 package lotto.controller
 
-import lotto.Constants
+import lotto.utils.Constants
 import lotto.model.LottoService
 import lotto.view.InputView
 import lotto.view.OutputView
 
 class LottoController(private val lottoService: LottoService) {
-    private val inputView = InputView()
-    private val outputView = OutputView()
 
     fun start() {
-        val lottoAmount = inputView.getLottoAmount()
+        val lottoAmount = InputView.getLottoAmount()
 
         val lottoCount = lottoAmount / Constants.LOTTO_PRICE
-        outputView.printLottoAmountMessage(lottoCount)
+        OutputView.printLottoAmountMessage(lottoCount)
 
         val lottoTickets = lottoService.generateLottoTickets(lottoCount)
-        outputView.printLottoNumber(lottoTickets)
+        OutputView.printLottoNumber(lottoTickets)
 
-        val winningNumbers = inputView.getWinningNumbers()
-        val bonusNumber = inputView.getBonusNumber(winningNumbers)
+        val winningNumbers = InputView.getWinningNumbers()
+        val bonusNumber = InputView.getBonusNumber(winningNumbers)
 
         val lottoRanks = lottoService.calculateMatchingNumbers(lottoTickets, winningNumbers, bonusNumber)
         val rankCount = lottoRanks.distinct().associateWith { rank -> lottoRanks.count { it == rank } }
 
-        outputView.printResult(rankCount)
+        OutputView.printResult(rankCount)
 
         val profit = lottoService.calculateProfit(rankCount, lottoAmount)
-        outputView.printProfit(profit)
+        OutputView.printProfit(profit)
     }
 
 }
